Integrate Linux/ASan memory-safety fixes (SlickAmogus#25 by Serentty)
Applies the portable fixes from the "More Linux work" PR to current code
(the PR branch predates our divergence, so applied by hand, not merged).
These are real latent bugs that just happen not to crash on Windows:

- libgs_stub.c: *(long*)&poly->x -> *(int*)&poly->x. `long` is 8 bytes on
  Linux LP64 so the screen-XY store overwrote the adjacent vertex; int is
  4 bytes on both ABIs. Byte-identical on Windows (long==int==32-bit).
- particle.c: gte_ldv0 reads 8 bytes but SVECTOR3 is 6 -> 2-byte stack
  over-read (func_800CE02C, Particle_SnowDraw). Use the 8-byte SVECTOR.
- data_stubs.c: D_800DF580 256->16*256 (twinfeeler/larva walks 16 elems);
  sharedData_800E21D0_0_s01 256->512 (AirScreamer struct is 0x170).
- Split globals the code walks across boundaries backed by one contiguous
  array + element aliases (ASan inter-global redzone trips): D_800A90A4
  descriptor lists (bodyprog.h/screen_data.c), D_8002AF04 waveform blob
  (bodyprog_80089090.c), inventory D_800C3E18[7]->[10] BSS aliasing +
  func_800548D8 light redirect (item_screens*.c).
- map7_s03_boss_motion.c/.py: sh_scr pool stride tracks sizeof (24B LP64 /
  12B ILP32). Byte-identical on Win64.
- CMake: on Linux, map .so's import PsyCross from the exe (single singleton
  instance) instead of linking their own copy; exe forces --whole-archive.
  Guarded UNIX-AND-NOT-APPLE / ELF-else; Windows path untouched.

All game-code changes are #ifdef SH_PC_PORT. Bumps PsyCross (PR#4 exports).
Verified: full Windows build (exe + 42 map DLLs) clean.

+++ pc_port/src/stubs/libgs_stub.c
@@ -400,7 +400,7 @@ void GsTMDfastF3LFG(void* op, VERT* vp, VERT* np, PACKET* pk, int n, int shift, 
         poly = (POLY_F3*)GsOUT_PACKET_P;
         setPolyF3(poly);
         setRGB0(poly, col_out.r, col_out.g, col_out.b);
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1; *(long*)&poly->x2 = sxy2;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1; *(int*)&poly->x2 = sxy2;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_F3);
     }
@@ -438,7 +438,7 @@ void GsTMDfastG3LFG(void* op, VERT* vp, VERT* np, PACKET* pk, int n, int shift, 
         setRGB0(poly, c0.r, c0.g, c0.b);
         setRGB1(poly, c1.r, c1.g, c1.b);
         setRGB2(poly, c2.r, c2.g, c2.b);
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1; *(long*)&poly->x2 = sxy2;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1; *(int*)&poly->x2 = sxy2;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_G3);
     }
@@ -474,8 +474,8 @@ void GsTMDfastF4LFG(void* op, VERT* vp, VERT* np, PACKET* pk, int n, int shift, 
         poly = (POLY_F4*)GsOUT_PACKET_P;
         setPolyF4(poly);
         setRGB0(poly, col_out.r, col_out.g, col_out.b);
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1;
-        *(long*)&poly->x2 = sxy2; *(long*)&poly->x3 = sxy3;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1;
+        *(int*)&poly->x2 = sxy2; *(int*)&poly->x3 = sxy3;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_F4);
     }
@@ -516,8 +516,8 @@ void GsTMDfastG4LFG(void* op, VERT* vp, VERT* np, PACKET* pk, int n, int shift, 
         setRGB2(poly, c2.r, c2.g, c2.b);
         /* v3 gets c2 color (PSX convention for quads) */
         setRGB3(poly, c2.r, c2.g, c2.b);
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1;
-        *(long*)&poly->x2 = sxy2; *(long*)&poly->x3 = sxy3;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1;
+        *(int*)&poly->x2 = sxy2; *(int*)&poly->x3 = sxy3;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_G4);
     }
@@ -557,7 +557,7 @@ void GsTMDfastTF3LFG(void* op, VERT* vp, VERT* np, PACKET* pk, int n, int shift,
         setUV3(poly, prim->tu0, prim->tv0, prim->tu1, prim->tv1, prim->tu2, prim->tv2);
         poly->tpage = prim->tpage;
         poly->clut  = prim->clut;
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1; *(long*)&poly->x2 = sxy2;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1; *(int*)&poly->x2 = sxy2;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_FT3);
     }
@@ -598,7 +598,7 @@ void GsTMDfastTG3LFG(void* op, VERT* vp, VERT* np, PACKET* pk, int n, int shift,
         setUV3(poly, prim->tu0, prim->tv0, prim->tu1, prim->tv1, prim->tu2, prim->tv2);
         poly->tpage = prim->tpage;
         poly->clut  = prim->clut;
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1; *(long*)&poly->x2 = sxy2;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1; *(int*)&poly->x2 = sxy2;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_GT3);
     }
@@ -639,8 +639,8 @@ void GsTMDfastTF4LFG(void* op, VERT* vp, VERT* np, PACKET* pk, int n, int shift,
                      prim->tu2, prim->tv2, prim->tu3, prim->tv3);
         poly->tpage = prim->tpage;
         poly->clut  = prim->clut;
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1;
-        *(long*)&poly->x2 = sxy2; *(long*)&poly->x3 = sxy3;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1;
+        *(int*)&poly->x2 = sxy2; *(int*)&poly->x3 = sxy3;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_FT4);
     }
@@ -684,8 +684,8 @@ void GsTMDfastTG4LFG(void* op, VERT* vp, VERT* np, PACKET* pk, int n, int shift,
                      prim->tu2, prim->tv2, prim->tu3, prim->tv3);
         poly->tpage = prim->tpage;
         poly->clut  = prim->clut;
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1;
-        *(long*)&poly->x2 = sxy2; *(long*)&poly->x3 = sxy3;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1;
+        *(int*)&poly->x2 = sxy2; *(int*)&poly->x3 = sxy3;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_GT4);
     }
@@ -715,7 +715,7 @@ void GsTMDfastNF3(void* op, VERT* vp, PACKET* pk, int n, int shift, GsOT* ot, un
         poly = (POLY_F3*)GsOUT_PACKET_P;
         setPolyF3(poly);
         setRGB0(poly, ITEMDIM(prim->r0), ITEMDIM(prim->g0), ITEMDIM(prim->b0));
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1; *(long*)&poly->x2 = sxy2;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1; *(int*)&poly->x2 = sxy2;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_F3);
     }
@@ -758,7 +758,7 @@ void GsTMDfastNG3(void* op, VERT* vp, PACKET* pk, int n, int shift, GsOT* ot, un
             poly = (POLY_F3*)GsOUT_PACKET_P;
             setPolyF3(poly);
             setRGB0(poly, ITEMDIM(gp->r0), ITEMDIM(gp->g0), ITEMDIM(gp->b0));
-            *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1; *(long*)&poly->x2 = sxy2;
+            *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1; *(int*)&poly->x2 = sxy2;
             ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
             GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_F3);
         }
@@ -783,7 +783,7 @@ void GsTMDfastNG3(void* op, VERT* vp, PACKET* pk, int n, int shift, GsOT* ot, un
         setRGB0(poly, ITEMDIM(prim->r0), ITEMDIM(prim->g0), ITEMDIM(prim->b0));
         setRGB1(poly, ITEMDIM(prim->r1), ITEMDIM(prim->g1), ITEMDIM(prim->b1));
         setRGB2(poly, ITEMDIM(prim->r2), ITEMDIM(prim->g2), ITEMDIM(prim->b2));
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1; *(long*)&poly->x2 = sxy2;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1; *(int*)&poly->x2 = sxy2;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_G3);
     }
@@ -814,8 +814,8 @@ void GsTMDfastNF4(void* op, VERT* vp, PACKET* pk, int n, int shift, GsOT* ot, un
         poly = (POLY_F4*)GsOUT_PACKET_P;
         setPolyF4(poly);
         setRGB0(poly, ITEMDIM(prim->r0), ITEMDIM(prim->g0), ITEMDIM(prim->b0));
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1;
-        *(long*)&poly->x2 = sxy2; *(long*)&poly->x3 = sxy3;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1;
+        *(int*)&poly->x2 = sxy2; *(int*)&poly->x3 = sxy3;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_F4);
     }
@@ -849,8 +849,8 @@ void GsTMDfastNG4(void* op, VERT* vp, PACKET* pk, int n, int shift, GsOT* ot, un
         setRGB1(poly, ITEMDIM(prim->r1), ITEMDIM(prim->g1), ITEMDIM(prim->b1));
         setRGB2(poly, ITEMDIM(prim->r2), ITEMDIM(prim->g2), ITEMDIM(prim->b2));
         setRGB3(poly, ITEMDIM(prim->r3), ITEMDIM(prim->g3), ITEMDIM(prim->b3));
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1;
-        *(long*)&poly->x2 = sxy2; *(long*)&poly->x3 = sxy3;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1;
+        *(int*)&poly->x2 = sxy2; *(int*)&poly->x3 = sxy3;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_G4);
     }
@@ -887,7 +887,7 @@ void GsTMDfastNTF3(void* op, VERT* vp, PACKET* pk, int n, int shift, GsOT* ot, u
         setUV3(poly, prim->tu0, prim->tv0, prim->tu1, prim->tv1, prim->tu2, prim->tv2);
         poly->tpage = prim->tpage;
         poly->clut  = prim->clut;
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1; *(long*)&poly->x2 = sxy2;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1; *(int*)&poly->x2 = sxy2;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_FT3);
     }
@@ -978,7 +978,7 @@ void GsTMDfastNTG3(void* op, VERT* vp, PACKET* pk, int n, int shift, GsOT* ot, u
         setUV3(poly, prim->tu0, prim->tv0, prim->tu1, prim->tv1, prim->tu2, prim->tv2);
         poly->tpage = prim->tpage;
         poly->clut  = prim->clut;
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1; *(long*)&poly->x2 = sxy2;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1; *(int*)&poly->x2 = sxy2;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_GT3);
     }
@@ -1014,8 +1014,8 @@ void GsTMDfastNTF4(void* op, VERT* vp, PACKET* pk, int n, int shift, GsOT* ot, u
                      prim->tu2, prim->tv2, prim->tu3, prim->tv3);
         poly->tpage = prim->tpage;
         poly->clut  = prim->clut;
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1;
-        *(long*)&poly->x2 = sxy2; *(long*)&poly->x3 = sxy3;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1;
+        *(int*)&poly->x2 = sxy2; *(int*)&poly->x3 = sxy3;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_FT4);
     }
@@ -1055,8 +1055,8 @@ void GsTMDfastNTG4(void* op, VERT* vp, PACKET* pk, int n, int shift, GsOT* ot, u
                      prim->tu2, prim->tv2, prim->tu3, prim->tv3);
         poly->tpage = prim->tpage;
         poly->clut  = prim->clut;
-        *(long*)&poly->x0 = sxy0; *(long*)&poly->x1 = sxy1;
-        *(long*)&poly->x2 = sxy2; *(long*)&poly->x3 = sxy3;
+        *(int*)&poly->x0 = sxy0; *(int*)&poly->x1 = sxy1;
+        *(int*)&poly->x2 = sxy2; *(int*)&poly->x3 = sxy3;
         ITEM_PRECISE_SZ(p); addPrim(&ot->org[otz], poly);
         GsOUT_PACKET_P = (u8*)poly + sizeof(POLY_GT4);
     }
